cicd/requirements.txt
cryptography
snowflake-connector-python
Jinja2
pyyaml